+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 * feature: 单元格事件
     - `props.cellHover`启用td hover状态
     - 新增 `cell-mouseenter` `cell-mouseleave` `cell-mouseover` 事件。
-* change: 减少css选择器层级。th background-color 变更为inherit
+* change: 点击行后，再次点击取消噶奥良。
+    - current-change 新增参数，{select: boolean} 表示是否选中。
+* change: 减少css选择器层级。th background-color 变更为inherit。
 * fix: 高亮行时，有概率不触发高亮的问题。
 
 ## 0.3.1